Create README.md
The name of our project is blood group management. Through this project we are doing, people will be given their name, mobile number and their blood group. If anyone is willing to donate blood, then if he writes his name, mobile and blood group, it will be saved and shown below in the form of table. If someone needs blood again, if he enters this website, he will be able to see his blood type and call him and collect his required blood from him. This project will help a person to get blood very fast.